requirements.txt
cloudscale-sdk
click>=7.0.0
tabulate
pygments
jmespath
yaspin
